requirements.txt
ib_insync
pandas
openai
tiktoken
flask
flask-cors
mysql-connector-python